sketchygame/sketchy_recogniser/pyproject.toml
[project]
name = "sketchy-recogniser"
version = "0.7.0"
description = "\"Cnn for sketches recognition\""
authors = [
    {name = "Jakub Gonczarek",email = "263999@student.pwr.edu.pl"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "path (>=17.1.0,<18.0.0)",
    "colorama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
